+++ pikoci/mysql/migrate/migrations/migrations.go
@@ -16,7 +16,7 @@ type Migration struct {
 // Migrations is the ordered list of all schema migrations. It is defined as
 // a fixed-size array so that the compiler catches ordering conflicts when
 // multiple developers add migrations concurrently.
-var Migrations = [30]Migration{
+var Migrations = [31]Migration{
 	V0Initial,
 	V1ResourceCheckInterval,
 	V2JobsAndBuilds,
@@ -47,4 +47,5 @@ var Migrations = [30]Migration{
 	V27ResourcePin,
 	V28JobTimeout,
 	V29SerialGroups,
+	V30WebhookTokenLength,
 }